Remove rsaPrivateKey
The current `rsaPrivateKey` is an array of four Big Numbers, where Big Number is defined by a Uint32Array-like array. The two first big numbers are the RSA key primes (p and q), the third is the private key exponent (d), the fourth is the Chinese remainder theorem q_inv.

As those are key primitives MEGA needs to recompute values of q1, p1, m, e, dp and dq in order to use that key. Because that's too complicated I think no one is using this value and I'm removing it from the public API.