Allow downloading midstream and the ciphertext
The start and end of stream can be changed with the
`start` and `end` options.

Downloading the ciphertext can be enabled using
`returnCiphertext: true` option.

For now `start` need to be a multiple of 16 (because AES
block size restrictions).

Also: improve error handling and disable chunking
when maxConnections is one.

Fixes #18 and #20.