Ciphertext upload WIP
Organized crypto/index.js moving MAC and key merging to a separate function.

Added `uploadCiphertext` option. The argument must be the key returned by `mega.encrypt`.

Not merged in master because if the `key` argument isn't provided it will generate other key and the download will be corrupt.

Probably the `uploadCiphertext` option will be removed and the `key` option will accept 32 byte key (current it only accepts 24 byte keys). In this case it will assume the data is already encrypted and the key is the one returned by `mega.encrypt`.