Fix and improve uploading, improve build process
The remaining buffer was being get from the beginning of the data
buffer, not from the end.

If the remaining buffer contained the entire final chunk it didn't
send it. Now it's fixed.

Off by one bugs fixed. We got one of the famous two hard things
in computer science!

Now using a single 'data' function instead of having one for the size
check and one for each chunk.

Now it's possible to filter which builds will be built.

This commit DON'T fixes login issues with new accounts neither removes
the Storage class on browser builds.

Fixes #44 (now that's for real!)