Fix some bugs
- Fix #46 by implementing support to new MEGA accounts
- Fix #45 by handling keys encrypted with multiple keys correctly
- Fix #32 by implementing qgustavor/mega-mock and using it to test the library (for now only logging-in, uploading and downloading are tested)

Also updated dependencies and implemented `forceHttps` argument for the File.prototype.download function.

Storage class is still built in browser versions. Logging-in into version 2 accounts on browsers now require WebCrypto support (and because of this HTTPS is required). This class might be removed in future.

As there isn't any breaking change a new minor version will be published.