I should have been commiting more often
The idea behind those changes is removing Babel dependencies by making Ava work using ES modules, but them I had to rename everything from .js to .mjs. I could add type: 'module' to package.json but them I think it could cause some issues to users of the library that were using it as a commonjs module.

Well, now the source code is a es module that can be run without any compiling. I wonder if would be better to just export the source code as the module build and just compile for cjs and browsers/deno. Better not: currently using esbuild allows to switch to typescript quite easily.

Browser tests removed: they were doing nothing important, just testing pure-javascript fallbacks, but currently the browser build is failing because the browser version of streams is not working well. Will be replaced with Deno based tests later as Deno is quite close to browsers but do not have the issue of being as slow and resource intensive as a browser.

Moved from Travis to Github Actions.