Add note about MAC implementation
Just a reminder to myself that implementing MEGA's weird MAC using native functions is not only hard, but makes the code run slower (at least on Node 16).

It's similar to this: https://en.wikipedia.org/wiki/CBC-MAC

But there is a small change on this that makes things harder: the computation result is added to a second ECBC-MAC and the main one resets every chunk.

Also, for some reason, trying to remove this.macs = [] by calculating the final mac in the checkMacBounding function made it slower too. The bad thing is that it increases memory usage, a 16 byte Buffer every 1 MB or so.

If someone wants to fix those things, do the benchmarks and send a pull request.