Handle callback, API and sharing issues
Setting userAgent to null will disable the user-agent header. At the moment it is needed in Firefox until Mega allows the user-agent header to be set in API requests.

User-agent and http agents are correctly accepted as API arguments

The unshare file method was implemented.

Logout callback is optional as before it didn't had a callback.