Handle upload progress better and don't minify browser-es
Upload progress can be monitored by checking progress events.
The browser-es build is no longer minified because I guess no one using it wants it to be minified, instead they would opt to use a minifier from a bundler which also would do tree shaking. Of they are using Deno, which, in this case, it's better not to minify to make debugging easier.